--- a/openquake/baselib/parallel.py
+++ b/openquake/baselib/parallel.py
@@ -747,7 +747,6 @@ class Starmap(object):
         self.todo = len(self.tasks)
         while self.todo:
             res = next(isocket)
-            assert self.calc_id, self.calc_id
             if self.calc_id != res.mon.calc_id:
                 logging.warning('Discarding a result from job %s, since this '
                                 'is job %d', res.mon.calc_id, self.calc_id)