--- a/openquake/calculators/getters.py
+++ b/openquake/calculators/getters.py
@@ -566,7 +566,6 @@ class RuptureGetter(object):
             sids = src_filter.close_sids(rup, self.trt, rup['mag'])
             weights.append(num_taxonomies_by_site[sids].sum())
         self.weights = numpy.array(weights)
-        self.weight = self.weights.sum()
 
     def split(self, maxweight):
         """