--- a/openquake/qa_tests_data/event_based_risk/case_miriam/job.ini
+++ b/openquake/qa_tests_data/event_based_risk/case_miriam/job.ini
@@ -41,7 +41,7 @@ minimum_magnitude = {"Subduction Interface": 5.5, "Active Shallow Crust": 5.2}
 # km
 maximum_distance = 200.0
 asset_hazard_distance = 20.0
-ebrisk_maxweight = 5E4
+ebrisk_maxweight = 10
 
 [event_based_params]