I am using now the same approach used in the classical calculator, i.e. the tasks are split depending on the estimated duration. Here the figures on cluster2 with @CatalinaYepes's calculation for Panama (ebr_Panama_City-occ.ini):

master: total time=8095s, slowest task=1,498s
this branch: total_time=2390s, slowest task=3,960s

The speedup is 3.7x.